+++ CHANGELOG.md
@@ -6,7 +6,8 @@
   a further check will be performed to see if the token has been revoked. 
 - [added] A new method ['auth.revoke_refresh_tokens(uid)'](https://firebase.google.com/docs/reference/admin/python/firebase_admin.auth#revoke_refresh_tokens)
   has been added to invalidate all tokens issued before the current second.
-- [added] A new property `tokens_valid_after_time` has been added to the ['UserRecord'](https://firebase.google.com/docs/reference/admin/python/firebase_admin.auth#userrecord)
+- [added] A new property `tokens_valid_after_time` has been added to the ['UserRecord'](https://firebase.google.com/docs/reference/admin/python/firebase_admin.auth#userrecord),
+  it denotes the time in epoch milliseconds before which tokens are not valid.
 
 # v2.8.0